[MOVE/#11] myPage 패키지 home에서 present 패키지 아래로 이동
