+++ app/src/main/java/com/puzzling/puzzlingaos/presentation/main/MainActivity.kt
@@ -7,7 +7,7 @@ import com.puzzling.puzzlingaos.R
 import com.puzzling.puzzlingaos.base.BaseActivity
 import com.puzzling.puzzlingaos.databinding.ActivityMainBinding
 import com.puzzling.puzzlingaos.presentation.home.HomeFragment
-import com.puzzling.puzzlingaos.presentation.home.mypage.MyPageFragment
+import com.puzzling.puzzlingaos.presentation.mypage.MyPageFragment
 
 class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
     override fun onCreate(savedInstanceState: Bundle?) {
