+++ app/src/main/java/com/puzzling/puzzlingaos/presentation/mypage/adapter/ChooseProjectAdapter.kt
@@ -1,4 +1,4 @@
-package com.puzzling.puzzlingaos.presentation.home.mypage.adapter
+package com.puzzling.puzzlingaos.presentation.mypage.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
