+++ app/src/main/res/layout/fragment_my_page.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.home.mypage.MyPageFragment">
+        tools:context=".presentation.mypage.MyPageFragment">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tb_myPage"
